Move NodeEnvironment to runtime folder
